% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addm.R
\name{addHOMA}
\alias{addHOMA}
\title{update an ogttCohort with HOMA as defined in 10.1210/jc.2002-021127}
\usage{
addHOMA(oc)
}
\arguments{
\item{oc}{ogttCohort instance}
}
\description{
update an ogttCohort with HOMA as defined in 10.1210/jc.2002-021127
}
\examples{
data(obaSamp)
addHOMA(obaSamp)
}
