% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addm.R
\name{addMinmodSIs}
\alias{addMinmodSIs}
\title{update an ogttCohort with insulin sensitivity by minimal model, with convergence information}
\usage{
addMinmodSIs(mae, iter = lapply, replace = TRUE, ...)
}
\arguments{
\item{mae}{instance of \code{\link{ogttCohort-class}}}

\item{iter}{lapply-like function, possibly mclapply}

\item{replace}{logical, if FALSE will fail if "SI" present among names of assays(mae)}

\item{\dots}{passed to \code{\link{getMinmodSIs}}}
}
\description{
update an ogttCohort with insulin sensitivity by minimal model, with convergence information
}
\examples{
data(obaSamp)
lk = addMinmodSIs(obaSamp[,1:5])
lk
}
