% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{concat}
\alias{concat}
\title{combine selected elements of a group of MultiAssayExperiments}
\usage{
concat(maes = list(base = omnicBase, CG = omnicCG, Cg = omnicCg, Cg = omnicCg,
  cg = omniccg), assays2keep = NULL, basenames = NULL, ...)
}
\arguments{
\item{maes}{a list of MultiAssayExperiment instances}

\item{assays2keep}{a character vector, used with subsetByAssay to restrict the elements of \code{maes}}

\item{basenames}{a character vector giving generic names for assay elements; if null, uses names of ElementList(maes)[[1]]}

\item{\dots}{not currently used}
}
\description{
combine selected elements of a group of MultiAssayExperiments
}
