% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bint3.R
\name{fig3plot}
\alias{fig3plot}
\title{plot analog of OMNICarb paper figure 3}
\usage{
fig3plot(f3out, measTag = "Matsuda")
}
\arguments{
\item{f3out}{output of fig3tests}

\item{measTag}{character tag to be used in xlab after delta symbol}
}
\description{
plot analog of OMNICarb paper figure 3
}
