% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmod.R
\name{fitOneMinMod}
\alias{fitOneMinMod}
\title{fit the dalla Man version of the Bergman minimal model (piecewise linear form for rate of appearance of glucose)}
\usage{
fitOneMinMod(insulin, glucose, bodyweight, times, nlsmaxit = 100,
  nlstol = 0.1, fullNLScontrol = NULL, nlstrace = FALSE, Sg = 0.028,
  V = 1.34, D = 75000, FA = 0.87, DC = 0.017, p2 = 0.012,
  nlsinit = list(a1 = 4, a2 = 4, a3 = 4, SI = 1e-04), trajTimes = 0:120,
  inputID = "")
}
\arguments{
\item{insulin}{a numeric vector of insulin values recorded at times in \code{times}}

\item{glucose}{a numeric vector of glucose concentrations, same length as \code{insulin}}

\item{bodyweight}{a numeric scalar}

\item{times}{times in minutes}

\item{nlsmaxit}{maximum number of iterations for NLS}

\item{nlstol}{tolerance for convergence of NLS}

\item{fullNLScontrol}{an optional list of control parameters for NLS}

\item{nlstrace}{logical requesting retention of trace}

\item{Sg}{fractional glucose effectiveness (default 0.028 min^-1)}

\item{V}{volume of distribution (default 1.34 dl/kg)}

\item{D}{ingested glucose dose in mg (default 75000)}

\item{FA}{fraction of glucose dose that is absorbed, default = 0.87, dimensionless}

\item{DC}{decay constant for model for rate of appearance of glucuse, default 0.017}

\item{p2}{rate constant of the remote insulin compartment from which insulin action emanates, default value 0.012 min^-1}

\item{nlsinit}{initial values for unknown parameters a1, a2, a3 (coefficients of piecewise linear model for rate of appearance of glucose) and SI (insulin sensitivity)}

\item{trajTimes}{a vector of times at which glucose and insulin trajectories should be estimated}

\item{inputID}{a string that may be used for rendering an ID tag}
}
\value{
a list with elements \code{fit} for the NLS fit, \code{soln} for the lsoda based solution to the differential equation model, \code{pwlinRa} for the piecewise-linear estimate of Ra (rate of appearance of glucose), \code{input} for a sublist with information about the input data, and \code{sessInf} for the sessionInfo metadata
}
\description{
fit the dalla Man version of the Bergman minimal model (piecewise linear form for rate of appearance of glucose)
}
