% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmod.R
\name{getMinmodSIs}
\alias{getMinmodSIs}
\title{Obtain SI from fits for all participants in an OGTT MultiAssayExperiment}
\usage{
getMinmodSIs(mae, iter = lapply, ...)
}
\arguments{
\item{mae}{MultiAssayExperiment instance}

\item{iter}{an lapply analog, supply mclapply for multicore (with options(mc.cores=...) appropriately set)}

\item{\dots}{passed to \code{\link{minmodByID}}}
}
\description{
Obtain SI from fits for all participants in an OGTT MultiAssayExperiment
}
