% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmod.R
\name{print.oneMinMod}
\alias{print.oneMinMod}
\alias{minmodByID}
\alias{print.minmodByID}
\title{Fit minimal model for a single subject}
\usage{
\method{print}{oneMinMod}(x, ...)

minmodByID(mae, id, gname = "glucose", iname = "insulin", ...)

\method{print}{minmodByID}(x, ...)
}
\arguments{
\item{\dots}{passed to \code{\link{fitOneMinMod}}}

\item{mae}{MultiAssayExperiment instance}

\item{id}{character id found in rownames(pData(mae))}

\item{gname}{name of ExperimentList component holding glucose concentrations}

\item{iname}{name of ExperimentList component holding insulin concentrations}
}
\description{
Fit minimal model for a single subject
}
\examples{
data(obaSamp)
m1 = minmodByID(obaSamp, "1")
m1
}
