% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCplot.R
\name{mvOutliers}
\alias{mvOutliers}
\title{get multivariate outlier indices separately for glucose and insulin series}
\usage{
mvOutliers(oc, t_glu = force, t_ins = force, ...)
}
\arguments{
\item{oc}{ogttCohort instance}

\item{t_glu}{function to transform glucose data to approximate multivariate normality}

\item{t_ins}{function to transform insulin data to approximate multivariate normality}

\item{\dots}{passed to \code{\link[parody]{mv.calout.detect}}}
}
\description{
get multivariate outlier indices separately for glucose and insulin series
}
\examples{
data(obaSamp)
mvOutliers(obaSamp)
mvOutliers(obaSamp, alpha=.1)
}
