% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obaSamp}
\alias{obaSamp}
\title{obaSamp: S4 container for baseline OGTT and related data}
\format{ogttCohort instance}
\source{
OMNICarb study, random sample of 50 subjects' baseline data
}
\usage{
obaSamp
}
\description{
obaSamp: S4 container for baseline OGTT and related data
}
\keyword{datasets}
