% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allS4.R
\docType{class}
\name{ogttCohort-class}
\alias{ogttCohort-class}
\title{an extension of MultiAssayExperiment including timing information}
\description{
an extension of MultiAssayExperiment including timing information
}
\section{Slots}{

\describe{
\item{\code{times}}{a numeric vector of OGTT sampling times}
}}

