% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omnicBase_samp}
\alias{omnicBase_samp}
\title{omniccg_samp: S4 container sample of tests at baseline}
\format{ogttCohort instance}
\source{
OMNICarb study
}
\usage{
omnicBase_samp
}
\description{
omniccg_samp: S4 container sample of tests at baseline
}
\keyword{datasets}
