% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omnicCg_samp}
\alias{omnicCg_samp}
\title{omnicCg_samp: S4 container sample of tests with diet Cg}
\format{ogttCohort instance}
\source{
OMNICarb study
}
\usage{
omnicCg_samp
}
\description{
omnicCg_samp: S4 container sample of tests with diet Cg
}
\keyword{datasets}
