% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot2}
\alias{plot2}
\title{crude visualization developed prior to plot_OGTT_fit}
\usage{
plot2(x, ...)
}
\arguments{
\item{x}{ogtt obj}

\item{\dots}{not used}
}
\description{
crude visualization developed prior to plot_OGTT_fit
}
