% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_OGTT_fit}
\alias{plot_OGTT_fit}
\title{produce a 4-panel plot of aspects of the OGTT minimal model fit}
\usage{
plot_OGTT_fit(x, ptitles = list(a = "(a)", b = "(b)", c = "(c)", d = "(d)"),
  ...)
}
\arguments{
\item{x}{output of minmodByID}

\item{ptitles}{a list of strings, with names 'a', ..., 'd', that will
be used as title elements for each panel}

\item{\dots}{not used}
}
\value{
northwest panel is a scatter plot of observed glucose vs. time,
with a solid line sketching the trajectory of the fitted mean glucose
based on the NLS fit; northeast panel is a scatterplot of observed insulin
vs. time; southwest panel is a sketch of fitted mean glucose against the
abstract variable X(t) [insulin action on glucose production and disposal];
southeast panel is a sketch of the fitted piecewise linear model for Ra(t),
the rate of appearance of glucose in the OGTT
}
\description{
produce a 4-panel plot of aspects of the OGTT minimal model fit
}
\examples{
data(obaSamp)
m1 = minmodByID(obaSamp, "1")
plot_OGTT_fit(m1)
}
