% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCplot.R
\name{QCplots}
\alias{QCplots}
\title{simple quality assessments for OGTT series in a cohort}
\usage{
QCplots(oc, choices = 1:2)
}
\arguments{
\item{oc}{ogttCohort instance}

\item{choices}{a 2-vector of integers passed as \code{choices} to \code{\link[ggbiplot]{ggbiplot}}, selecting principal axes for biplot display}
}
\description{
simple quality assessments for OGTT series in a cohort
}
\examples{
example(csvImport) # makes democ
QCplots(democ)
}
