% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SItests}
\alias{SItests}
\title{SItests: cached set of pairwise contrasts for minimal model-based SI (insulin sensitivity) estimates}
\format{list}
\source{
OMNICarb study
}
\usage{
SItests
}
\description{
SItests: cached set of pairwise contrasts for minimal model-based SI (insulin sensitivity) estimates
}
\keyword{datasets}
