% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xoverSamp}
\alias{xoverSamp}
\title{xoverSamp: S4 container multiple diet data sample}
\format{ogttCohort instance}
\source{
OMNICarb study
}
\usage{
xoverSamp
}
\description{
xoverSamp: S4 container multiple diet data sample
}
\keyword{datasets}
