% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsuda.R
\name{addMatsuda120}
\alias{addMatsuda120}
\title{Compute Matsuda's index (120 minute form) for all samples}
\usage{
addMatsuda120(mae, gname = "glucose", iname = "insulin",
  outname = "Mats120", allowGaps = FALSE)
}
\arguments{
\item{mae}{instance of \code{\link[MultiAssayExperiment]{MultiAssayExperiment-class}}}

\item{gname}{name of ExperimentList component holding glucose concentrations}

\item{iname}{name of ExperimentList component holding insulin concentrations}

\item{outname}{name of ExperimentList component holding Matsuda index}

\item{allowGaps}{logical, should addAssay fail if there are gaps between sample labels for new assay and rownames of colData}
}
\value{
instance of \code{\link[MultiAssayExperiment]{MultiAssayExperiment-class}} that includes assay \code{"Mats120"}
}
\description{
Compute Matsuda's index (120 minute form) for all samples
}
\note{
The formula of the WEB CALCULATOR at \url{http://mmatsuda.diabetes-smc.jp/english.html} is used
}
